Extracted FoodData to seperate module
