+++ src/tests/hoplalib/zoo/test_petmodels.py
@@ -6,7 +6,7 @@ import pytest
 
 from hopla.cli.groupcmds.get_user import HabiticaUser
 from hopla.hoplalib.common import GlobalConstants
-from hopla.hoplalib.zoo.foodmodels import FoodData
+from hopla.hoplalib.zoo.fooddata import FoodData
 from hopla.hoplalib.zoo.petdata import PetData
 from hopla.hoplalib.zoo.petmodels import FeedingStatus, InvalidFeedingStatus, \
     InvalidPet, Pet, PetMountPair, Zoo, ZooBuilder
